pesca_backend/requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
asgiref==3.11.1
astunparse==1.6.3
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
dj-database-url==3.1.2
Django==6.0.2
django-storages==1.14.6
djangorestframework==3.16.1
flatbuffers==25.12.19
frozenlist==1.8.0
gast==0.7.0
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-cloud-vision==3.12.1
google-crc32c==1.8.0
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h5py==3.14.0
http_ece==1.2.1
idna==3.11
keras==3.13.2
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
mercadopago==2.3.0
ml_dtypes==0.5.4
multidict==6.7.1
namex==0.1.0
numpy==2.4.3
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
py-vapid==1.9.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
pytesseract==0.3.13
python-dotenv==1.2.2
pywebpush==2.3.0
qrcode==8.2
requests==2.32.5
rich==14.3.3
rsa==4.9.1
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
tensorflow==2.21.0
termcolor==3.3.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wheel==0.46.3
whitenoise==6.12.0
wrapt==2.1.2
yarl==1.23.0
